refactor(upstream): extract prompts_list.rs + prompts_get.rs (lab-kvji.12.8 step 17)
- prompts_list.rs: collect_upstream_prompts, list_upstream_prompts,
  cached_upstream_prompt_names, cached_prompt_owner, prompt_ownership_map,
  find_prompt_owner + merge/successful_prompt/2 owner_lookup tests.
  cached_prompt_owner + find_prompt_owner co-located (no flip, §2.1 drop note).
- prompts_get.rs: subject_scoped_prompts, subject_scoped_prompt_owner, get_prompt,
  subject_scoped_get_prompt + get_prompt_times_out test.
NO-TOUCH verified: all 3 subject_scoped_* prompt methods retain subject threading.
pool.rs test mod fully drained and removed; pool.rs now 160 LOC (coordinator).
Test count 1423, no test lost (43==43 vs baseline).

Bead: lab-kvji.12.8
